+++ adapter/cd/src/cd/config.rs
@@ -1,0 +1,3 @@
+pub struct Config {
+    pub socket_path: String,
+}
