+++ adapter/cd/src/main.rs
@@ -1,44 +1,39 @@
-
-use std::env;
-use std::io::prelude::*;
-use std::io;
-use std::path::Path;
-use std::fs::OpenOptions;
-use std::sync::Arc;
 use daemonize::Daemonize;
-
+use std::env;
+use std::fs::OpenOptions;
+use std::io;
+use std::io::prelude::*;
+use std::path::Path;
+use std::sync::Arc;
 
 pub mod cd;
 
-
-
 const LOG_DIR: &str = "/var/run/zpr";
 const PID_DIR: &str = "/var/run/zpr";
-
-
 
 fn usage() -> io::Result<()> {
     println!("Usage: cd [-f|--foreground]");
     println!("ZPR Connection Daemon\n");
     println!("Unless the foreground option is provided, the daemon will run in the");
-    println!("background, logging to {}/cd.out and {}/cd.err.", LOG_DIR, LOG_DIR);
+    println!(
+        "background, logging to {}/cd.out and {}/cd.err.",
+        LOG_DIR, LOG_DIR
+    );
     println!();
 
     Ok(())
 }
 
-
 fn main() -> io::Result<()> {
-    let args: Vec<String> = env::args().collect();    
+    let args: Vec<String> = env::args().collect();
     if args.contains(&String::from("-h")) || args.contains(&String::from("--help")) {
         return usage();
     }
     let foreground = if args.len() > 1 {
-        args[1] == "-f" || args[1] == "--foreground"            
-        } else {
-            false
-        };  
-
+        args[1] == "-f" || args[1] == "--foreground"
+    } else {
+        false
+    };
 
     let config = Arc::new(cd::Config {
         socket_path: String::from("/var/run/zpr/cd.sock"),
@@ -46,10 +41,9 @@ fn main() -> io::Result<()> {
 
     if foreground {
         return cd::tokio_main(config.clone());
-    }    
+    }
 
     // Else we go into background.
-
 
     let logpath = Path::new(LOG_DIR);
     if !logpath.exists() {
@@ -65,13 +59,21 @@ fn main() -> io::Result<()> {
         .append(true)
         .truncate(false)
         .open(format!("{}/cd.out", LOG_DIR))?;
-    write!(stdout, "=============== cd restarts at {} ==============\n", chrono::Local::now())?;
+    write!(
+        stdout,
+        "=============== cd restarts at {} ==============\n",
+        chrono::Local::now()
+    )?;
     let mut stderr = OpenOptions::new()
         .create(true)
         .append(true)
         .truncate(false)
         .open(format!("{}/cd.err", LOG_DIR))?;
-    write!(stderr, "=============== cd restarts at {} ==============\n", chrono::Local::now())?;    
+    write!(
+        stderr,
+        "=============== cd restarts at {} ==============\n",
+        chrono::Local::now()
+    )?;
 
     let daemonize = Daemonize::new()
         .pid_file(format!("{}/cd.pid", PID_DIR))
@@ -84,6 +86,3 @@ fn main() -> io::Result<()> {
     }
     cd::tokio_main(config.clone())
 }
-
-
-
